refactor: rename 'budget' to 'price' and restructure milestone UI
- renamed all 'budget' references to 'price' across Entrust Upload files
- moved ArtworkEntrustSetTab.jsx/css to components/Tabs directory
- renamed Dropdown.css to EntrustUploadFormDropdown.module.css
- improved mobile responsiveness of milestone layout in EntrustUploadForm5
